fix: keep rtmp alive and allow new conenctions
